+++ lib/irb/cmd/edit.rb
@@ -1,5 +1,6 @@
 require 'shellwords'
 require_relative "nop"
+require_relative "../source_finder"
 
 module IRB
   # :stopdoc:
@@ -28,11 +29,9 @@ module IRB
         end
 
         if !File.exist?(path)
-          require_relative "show_source"
-
           source =
             begin
-              ShowSource.find_source(path, @irb_context)
+              SourceFinder.new(@irb_context).find_source(path)
             rescue NameError
               # if user enters a path that doesn't exist, it'll cause NameError when passed here because find_source would try to evaluate it as well
               # in this case, we should just ignore the error
